[core] fix crash when using lighttpd -1 with pipes (fixes #3117)
(thx povcfe-bug)

In some scenarios with lighttpd -1 on pipes, the event handler
associated with the connection (con) will have been cleaned up,
and a pending event on the one-off oneshot_fdn will not yet have
been handled, leading to a crash.  lighttpd typically is used to
handle requests on sockets, not pipes.  With nc (netcat), the
requests might be sent over pipes.  The bug occurred with the
accommodations made in lighttpd to handle the request over pipes.

x-ref:
  "Trigger crash when using lighttpd -1 with pipes"
  https://redmine.lighttpd.net/issues/3117