[mod_magnet] lighty.c.header_tokens convenience fn
lighty.c.header_tokens convenience func to create a sequence table
of tokens parsed from a given string, e.g. an HTTP header.  The purpose
of this routine is to made it easier to properly parse an HTTP header
into tokens since token separators can be part of quoted-strings, and
they are not token separators when part of quoted strings.

The sequence table t returned from lighty.c.header_tokens() can be
walked with:
  for i = 1, #t do
    -- <body here>
  end
While walking, each element can be passed to lighty.c.quoteddec()
to decode, as non-quoted-string elements are returned as-is.

Note: lighty.c.header_tokens() returns a sequence table,
      which is different from lighty.c.cookie_tokens(),
      which returns a key/value table of cookies.