[build] improve searching for PostgreSQL
Use CMake builtin module to search for PostgreSQL if a recent enough version of
CMake is used. Properly look for the include otherwise, it may be in a directory
that is not in the compiler default search paths.

There is no need to look for pg_config, the returned values were not used at
all. It also returns information on the server components, which are of no real
interest here, so the script may not even be present even if the client
development headers are installed.