[core] build core modules into lighttpd executable
In dynamic build of lighttpd, at runtime each loaded lighttpd module .so
takes at least (5) 4k memory pages on x86_64.  That is at least 20k per
module.  Building 10 small core modules into lighttpd adds less than 16k
.text total, and adding mod_fastcgi and mod_scgi totals less than 20k
.text (plus an extra 4k .rodata page).  The end result is that memory
usage is almost the same if only mod_staticfile is used, and 20k or so
is saved for each additional module used if among those built into
lighttpd executable.

For now, the standalone .so modules continue to be built, even though
not used by the lighttpd executable.  This is done in order to give
distros time to update packaging scripts and transition to not
expecting the builtin module artifacts to be produced or installed as
part of the distro lighttpd package.