Added masking options for remote address logging
This change is inspired by mod_log_ipmask for the apache webserver,
which was originally developed by the data protection office in
Saxonia, Germany (see
https://www.saechsdsb.de/component/jifile/download/ZjU5MGFlZjU5MzczYzdhOTNkM2NhYjFiYzg1OGU0ZWE=/dokumentation-ip-maskierung-v1-0-pdf ).
The goal is to de-personalize the IP address, which is otherwise
considered a piece of personally identifiable information (in some
jurisdictions at least).

The code here was written from scratch by myself and is not a work
derived from the original mod_log_ipmask code.

It adds an optional parameter of the form "[v4|v6]:n[,...]" to the
"%a" / "%h" placeholders in lighttpd's accesslog.format configuration
parameter, where n is the number of trailing bits to mask out in
addresses belonging to the respective protocols, for the purpose of
logging. For example, the placeholder "%{v4:8,v6:72}a" would replace
the rightmost 8 bits of IPv4 addresses with 0's, and the rightmost
72 bits of IPv6 addresses. In the absence of the parameter the old
behaviour is preserved, i. e. the full address is logged.