[core] always decr fd count upon socket close()
always decr fd count in connection_close()
always decr fd count in fdevent_sched_run()

Error return value from close() should be used for diagnostics and
recovery, but the state of the file descriptor is unspecified by POSIX.
On most systems, it is invalid to redo close().  (Linux 'man 2 close'
suggests that HP-UX is an outlier, and that a future POSIX standard
update will specify the behavior for the file descriptor to be closed)

EBADF should not happen in those funcs for lighttpd since those should
be the only locations in lighttpd where those fds are closed.