[core] allow larger number of Ranges if sorted
Protocols like zchunk and zsync may try to send a large number of Ranges
in a single HTTP request.  Track if ranges are sorted and allow larger
number of ranges if the list of ranges is already sorted ascending.

If unsorted ranges are encountered, the limit remains 10 ranges due to
the non-linear cost to detect overlapping ranges or other range attacks.