[mod_dirlisting] dir-listing.sort option (#3235)
dir-listing.sort = "enable"  # (default)

option for server to produce unsorted output for directory listing
(may be useful on embedded systems listing large folders on SD cards)

Note: clients which support javascript will continue to sort the listing
on the client side when using the auto layout and javascript defaults
provided by mod_dirlisting.

This option allows for a streamable response (optional) and avoids
memory used to read the entire directory and produce a sorted listing,
as memory used can add up for large folders, especially if temporary
files for response are configured to be on in-memory tmpfs filesystems.

Admin should (separately) enable server.stream-response-body=1 or =2
in lighttpd.conf to enable streaming of the response.  Alternatively,
leave the defaults and use mod_deflate to compress the response for
clients which send Accept-Encoding.

Note: if dir-listing.cache is configured, then cache should be cleared
if dir-listing.sort is changed in config from disabled to enabled
(unless dir-listing.cache max-age is sufficiently short that it will
 sort itself out momentarily (pun intended)).

x-ref:
  "mod_dirlisting: embedded device unresponsive when accessing excessively large folders"
  https://redmine.lighttpd.net/issues/3235