[core] cfg "if","elif","elsif","elseif","else if"
accept "if", "elif", "elsif", "elseif", "else if" in config
These tokens are more intuitive syntax and correspond to other langs.

Using the historical 'else <condition>' still works.

This patch does not prevent replacing a final else { ... } condition
with elseif { ... }, but users should not skip the condition when using
"elif", "elsif", "elseif", or "else if" as doing so may become a syntax
error in the future.