[scons] remove CheckFunc() incorrect header usage
From the SCons doc for CheckFunc():
Note: if header is supplied, it should not include the standard header file that declares function_name, and it should include a dummy prototype similar to the default case. Compilers reject builds where a function call does not match the declared prototype as happens if the "real" header is included, and modern compilers are now rejecting implicit function declarations.