[mod_magnet] lighty.r.req_body.unspecified_len
lighty.r.req_body.unspecified_len() - treat HTTP/1.0 requests without
specifying Content-Length and without keep-alive to stream a request
body of unspecified length.  The default in lighttpd is to treat such
HTTP/1.0 requests as if Content-Length: 0 was sent (e.g. for GET
requests, which do not require a Content-Length to be set in HTTP/1.0).

x-ref:
  "live555 rtp/rtsp over http reverse proxy 1.4.73"
  https://redmine.lighttpd.net/boards/2/topics/11397