[core] add default to builtin mimetype.assign
add default "application/octet-stream" to builtin mimetype.assign

If mimetype.assign is not specified in lighttpd.conf, the builtin
default for mimetype.assign is used.  This change adds a default
  "" => "application/octet-stream"
to the builtin mimetype.assign.  The effect is that ETag and
Last-Modified are sent for static files which use this default,
where previously ETag and Last-Modified were omitted for static
files without a specified Content-Type.  This aligns with widely used
behavior in distro packages of lighttpd, which specify mimetype.assign
to include "" => "application/octet-stream"

The historical reason for omitting ETag and Last-Modified for static
files with unspecified Content-Type was to avoid first-time user
misconfiguration of lighttpd.conf, resulting in download of files
intended to be executed by CGI (e.g. .php, .pl, .py), and then
over-aggressive caching by browsers even after the misconfiguration
was corrected.  The over-agressive caching by browsers in this case
seems to be less of an issue today than it was, and many distro packages
additionally include lighttpd.conf protection against downloading code,
  e.g. static-file.exclude-extensions = ( ".php", ".pl", ".py" )