[core] preprocessor option to force crypto lib
add preprocessor option to force crypto lib used by lighttpd base
  -DFORCE_OPENSSL_CRYPTO for OpenSSL
  or -DFORCE_GNUTLS_CRYPTO for GnuTLS
in addition to the already-existing
  -DFORCE_MBEDTLS_CRYPTO for mbed TLS
  or -DFORCE_WOLFSSL_CRYPTO for WolfSSL

This allows building lighttpd with support for multiple TLS modules,
and at the same time specifying the lighttpd base crypto lib dependency.

(Packages such as OpenWRT lighttpd package would need to update
package dependency along with specifying the preprocessor define)

If lighttpd is built with meson and without "-Dwith_nettle=true",
then lighttpd will use mbed TLS if built with "-Dwith_mbedtls",
so the preprocessor option is not strictly necessary for mbed TLS.
However, if built with "-Dwith_mbedtls" and "-Dwith_wolfssl" *and*
without "-Dwith_nettle=true" *and* lighttpd base crypto dependency
on WolfSSL is desired, then CFLAGS -DFORCE_WOLFSSL_CRYPTO is needed,
e.g. with a patch to add it to 'defs' near the top of src/meson.build.
The defines -DFORCE_OPENSSL_CRYPTO or -DFORCE_GNUTLS_CRYPTO are similar
to -DFORCE_WOLFSSL_CRYPTO in that they must be specified to have the
desired effect.

If any -DFORCE_*_CRYPTO is defined, then lighttpd will not use Nettle
and *should not* be built with "-Dwith_nettle=true", in order to avoid
an unnecessary link dependency (though also handled in meson.build).

x-ref:
  "lighttpd authentication modules depends on libnettle"
  https://github.com/openwrt/packages/issues/18851
  "lighttpd: fix missing support for OpenSSL crypto library"
  https://github.com/openwrt/packages/pull/24004