[mod_dirlisting] sort mtime using data-value (#3264)
date is parsed at runtime, upon selection of date column for sorting,
and the data-value is calculated and saved for future sort actions.

(data-value could be written into HTML during page generation, but would
result in larger HTML page size where the additional data used only for
sorting the date column.  Therefore, the choice has been made to send a
smaller HTML page and defer the work to parse the dates to the initial
sorting of the date column, which might not occur for a given client.)

x-ref:
  https://redmine.lighttpd.net/issues/3264
  "[PATCH] sort by the exact value of mtime or size"