[mod_proxy] sketch out streaming and trailers
sketch out (not enabled) streaming and trailers to mod_proxy backend
(when mod_proxy sends HTTP/1.1 Transfer-Encoding: chunked to backend)

Note: _L_TRAILERS lighttpd env addition would be too late for trailers
to be added to backend environments like FastCGI, SCGI, AJP13, etc.
We already merge trailers into headers if trailers are fully received
before connect to backend, e.g. if not streaming request body to backend

Trailers are not widely used, and so this sketched code is disabled.
Before enabling the code in this code, also consider memory use to
accumulate trailers and hold them for the duration of the request
and response when request is not handled by mod_proxy or mod_proxy
is not sending Transfer-Encoding: chunked to backend.  (This could be
mitigated with a dedicated flag or other options.)