[core] security: fix to reject disallowed trailers
(thx sebsrt)

lighttpd mod_proxy sends Connection: close to backends, so this bug is
not exploitable to send additional requests to backends, as long as the
backend properly closes the connection after seeing Connection: close
in the request lighttpd sends to the backend.

However, trailers could (incorrectly) modify the request headers sent to
the backend which might bypass configured policy which checked request
headers.

bug was introduced in lighttpd 1.4.80 and is fixed in lighttpd 1.4.81